Move part of src/space to hyperon-space lib.
