+++ hyperon-space/src/index/mod.rs
@@ -220,9 +220,8 @@ mod test {
     use super::*;
 
     use hyperon_atom::matcher::*;
-    use hyperon_atom::{expr, sym, bind};
+    use hyperon_atom::{expr, sym, bind, number::Number};
     use hyperon_common::assert_eq_no_order;
-    use crate::metta::runner::number::Number;
     use std::fmt::{Debug, Display, Formatter};
 
     #[test]
