+++ hyperon-space/src/index/trie.rs
