+++ lib/benches/type.rs
@@ -2,10 +2,10 @@
 
 extern crate test;
 
+use hyperon_atom::Atom;
 use test::Bencher;
-use hyperon::*;
 use hyperon::metta::types::*;
-use hyperon::space::DynSpace;
+use hyperon_space::DynSpace;
 use hyperon::space::grounding::*;
 use hyperon::metta::text::*;
