+++ lib/src/metta/runner/mod.rs
@@ -60,7 +60,7 @@ use hyperon_atom::*;
 use hyperon_common::shared::Shared;
 
 use super::*;
-use crate::space::*;
+use hyperon_space::*;
 use super::text::{Tokenizer, Parser, SExprParser};
 use super::types::validate_atom;
 
@@ -93,10 +93,6 @@ use stdlib::CoreLibLoader;
 
 mod builtin_mods;
 use builtin_mods::*;
-
-pub mod bool;
-pub mod number;
-pub mod str;
 
 const EXEC_SYMBOL : Atom = sym!("!");
