+++ lib/src/metta/runner/stdlib/atom.rs
@@ -1,12 +1,12 @@
 use hyperon_atom::*;
-use crate::space::*;
+use hyperon_space::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use crate::metta::types::{get_atom_types, get_meta_type};
 use hyperon_common::multitrie::{MultiTrie, TrieKey, TrieToken};
 use hyperon_common::collections::ImmutableString;
 use super::{grounded_op, regex};
-use crate::metta::runner::number::*;
+use hyperon_atom::number::*;
 
 use std::convert::TryInto;
 use std::hash::{DefaultHasher, Hasher};
