+++ lib/src/metta/runner/stdlib/debug.rs
@@ -5,8 +5,8 @@ use hyperon_common::collections::{SliceDisplay, Equality, DefaultEquality};
 use hyperon_common::assert::compare_vec_no_order;
 use hyperon_atom::matcher::atoms_are_equivalent;
 use crate::metta::runner::stdlib::{grounded_op, regex, unit_result};
-use crate::metta::runner::bool::*;
-use crate::metta::runner::str::*;
+use hyperon_atom::bool::*;
+use hyperon_atom::str::*;
 use hyperon_atom::gnd::GroundedFunctionAtom;
 
 use std::convert::TryInto;
