+++ lib/src/metta/runner/stdlib/math.rs
@@ -2,8 +2,8 @@ use hyperon_atom::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use super::{grounded_op, regex};
-use crate::metta::runner::number::*;
-use crate::metta::runner::bool::*;
+use hyperon_atom::number::*;
+use hyperon_atom::bool::*;
 
 use std::convert::TryInto;
