+++ lib/src/metta/runner/stdlib/module.rs
@@ -1,11 +1,11 @@
 use hyperon_atom::*;
-use crate::space::*;
+use hyperon_space::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use hyperon_common::shared::Shared;
 use crate::metta::runner::{Metta, RunContext, ResourceKey};
 use super::{grounded_op, regex, unit_result};
-use crate::metta::runner::str::expect_string_like_atom;
+use hyperon_atom::str::expect_string_like_atom;
 
 use regex::Regex;
