+++ lib/src/metta/runner/stdlib/space.rs
@@ -1,5 +1,5 @@
 use hyperon_atom::*;
-use crate::space::*;
+use hyperon_space::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
 use crate::metta::runner::stdlib::{grounded_op, unit_result, regex};
